Finishes migration to rocket
